# backend/openapi.json
{
  "components": {
    "schemas": {
      "Cuisine": {
        "description": "Fixed list; every recipe has exactly one. \"other\" covers anything not listed.",
        "enum": [
          "british",
          "chinese",
          "indian",
          "italian",
          "mediterranean",
          "mexican",
          "thai",
          "other"
        ],
        "title": "Cuisine",
        "type": "string"
      },
      "DietaryLabel": {
        "description": "Fixed list used for filtering. A recipe has zero or more.",
        "enum": [
          "vegetarian",
          "vegan",
          "gluten-free",
          "dairy-free"
        ],
        "title": "DietaryLabel",
        "type": "string"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "IngredientLine": {
        "properties": {
          "ingredientId": {
            "title": "Ingredientid",
            "type": "integer"
          },
          "item": {
            "title": "Item",
            "type": "string"
          },
          "prep": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "ingredientId",
          "item",
          "quantity",
          "unit"
        ],
        "title": "IngredientLine",
        "type": "object"
      },
      "IngredientLineIn": {
        "properties": {
          "item": {
            "minLength": 1,
            "title": "Item",
            "type": "string"
          },
          "prep": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "item"
        ],
        "title": "IngredientLineIn",
        "type": "object"
      },
...
      "MealType": {
        "description": "What a recipe suits. A recipe has one or more; used as a filter (the plan has no fixed slots).",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "dessert"
        ],
        "title": "MealType",
        "type": "string"
      },
...
      "RecipeCreate": {
        "description": "POST /api/recipes and PUT /api/recipes/{id} (custom recipes only).",
        "properties": {
          "cuisine": {
            "$ref": "#/components/schemas/Cuisine"
          },
          "dietary": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DietaryLabel"
            },
            "title": "Dietary",
            "type": "array"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientLineIn"
            },
            "minItems": 1,
            "title": "Ingredients",
            "type": "array"
          },
          "mealTypes": {
            "items": {
              "$ref": "#/components/schemas/MealType"
            },
            "minItems": 1,
            "title": "Mealtypes",
            "type": "array"
          },
          "method": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Method",
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "serves": {
            "minimum": 1.0,
            "title": "Serves",
            "type": "integer"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "name",
          "cuisine",
          "serves",
          "mealTypes",
          "ingredients",
          "method"
        ],
        "title": "RecipeCreate",
        "type": "object"
      },
      "RecipeDetail": {
        "description": "GET /api/recipes/{id}?servings=N. Quantities are scaled when servings is given.",
        "properties": {
          "cuisine": {
            "$ref": "#/components/schemas/Cuisine"
          },
          "defaultImage": {
            "title": "Defaultimage",
            "type": "string"
          },
          "dietary": {
            "items": {
              "$ref": "#/components/schemas/DietaryLabel"
            },
            "title": "Dietary",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientLine"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "isCustom": {
            "title": "Iscustom",
            "type": "boolean"
          },
          "mealTypes": {
            "items": {
              "$ref": "#/components/schemas/MealType"
            },
            "title": "Mealtypes",
            "type": "array"
          },
          "method": {
            "items": {
              "type": "string"
            },
            "title": "Method",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "serves": {
            "title": "Serves",
            "type": "integer"
          },
          "servings": {
            "title": "Servings",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "cuisine",
          "serves",
          "mealTypes",
          "dietary",
          "tags",
          "isCustom",
          "imageUrl",
          "defaultImage",
          "servings",
          "ingredients",
          "method"
        ],
        "title": "RecipeDetail",
        "type": "object"
      },
      "RecipeSummary": {
        "description": "GET /api/recipes item.",
        "properties": {
          "cuisine": {
            "$ref": "#/components/schemas/Cuisine"
          },
          "defaultImage": {
            "title": "Defaultimage",
            "type": "string"
          },
          "dietary": {
            "items": {
              "$ref": "#/components/schemas/DietaryLabel"
            },
            "title": "Dietary",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "isCustom": {
            "title": "Iscustom",
            "type": "boolean"
          },
          "mealTypes": {
            "items": {
              "$ref": "#/components/schemas/MealType"
            },
            "title": "Mealtypes",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "serves": {
            "title": "Serves",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "cuisine",
          "serves",
          "mealTypes",
          "dietary",
          "tags",
          "isCustom",
          "imageUrl",
          "defaultImage"
        ],
        "title": "RecipeSummary",
        "type": "object"
      },
      "RecipeUsageOut": {
        "description": "ASSUMPTION (3): GET /api/recipes/{id}/usage -> {\"upcomingMeals\": N} for the delete confirmation.",
        "properties": {
          "upcomingMeals": {
            "title": "Upcomingmeals",
            "type": "integer"
          }
        },
        "required": [
          "upcomingMeals"
        ],
        "title": "RecipeUsageOut",
        "type": "object"
      },
...
      "TagOut": {
        "properties": {
          "isBuiltin": {
            "title": "Isbuiltin",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "isBuiltin"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Meal planning for households on a tight budget. Errors always use the `ErrorResponse` object.",
    "title": "Nosh API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/recipes": {
      "get": {
        "description": "Search and filter recipes. Dietary filter defaults to the saved preferences unless `dietary` is given or `all=true`. Dietary labels must **all** match (vegetarian also accepts vegan); meal types and tags match **any**. Search covers recipe names and ingredients.",
        "operationId": "list_recipes",
        "parameters": [
          {
            "description": "Search text (name or ingredient)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search text (name or ingredient)",
              "title": "Q"
            }
          },
          {
            "description": "Comma-separated, e.g. `lunch,dinner`",
            "in": "query",
            "name": "mealType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated, e.g. `lunch,dinner`",
              "title": "Mealtype"
            }
          },
          {
            "description": "Comma-separated, e.g. `vegetarian,gluten-free`",
            "in": "query",
            "name": "dietary",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated, e.g. `vegetarian,gluten-free`",
              "title": "Dietary"
            }
          },
          {
            "description": "Comma-separated tag keys, e.g. `quick,low-cost`",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated tag keys, e.g. `quick,low-cost`",
              "title": "Tag"
            }
          },
          {
            "description": "Ignore saved dietary preferences",
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "description": "Ignore saved dietary preferences",
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeSummary"
                  },
                  "title": "Response List Recipes",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "List recipes",
        "tags": [
          "Recipes"
        ]
      },
      "post": {
        "description": "Adds a custom recipe. Unknown tag names are created. The slug is made from the name.",
        "operationId": "create_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict (e.g. a recipe with this name already exists)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Create a recipe",
        "tags": [
          "Recipes"
        ]
      }
    },
    "/api/recipes/{slug}": {
      "delete": {
        "description": "Soft-deletes a custom recipe: removed from today and future days, kept in past weeks.",
        "operationId": "delete_recipe",
        "parameters": [
          {
            "description": "Recipe slug, e.g. `lentil-dahl`",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Recipe slug, e.g. `lentil-dahl`",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed (e.g. editing a built-in recipe)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Delete a recipe",
        "tags": [
          "Recipes"
        ]
      },
      "get": {
        "description": "Full recipe. Pass `servings` to scale ingredient amounts.",
        "operationId": "get_recipe",
        "parameters": [
          {
            "description": "Recipe slug, e.g. `lentil-dahl`",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Recipe slug, e.g. `lentil-dahl`",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Scale amounts to this many servings",
            "in": "query",
            "name": "servings",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 50,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scale amounts to this many servings",
              "title": "Servings"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Get a recipe",
        "tags": [
          "Recipes"
        ]
      },
      "put": {
        "description": "Replaces a custom recipe. Built-in recipes are read-only. Renaming updates the slug.",
        "operationId": "update_recipe",
        "parameters": [
          {
            "description": "Recipe slug, e.g. `lentil-dahl`",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Recipe slug, e.g. `lentil-dahl`",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed (e.g. editing a built-in recipe)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict (e.g. a recipe with this name already exists)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Update a recipe",
        "tags": [
          "Recipes"
        ]
      }
    },
    "/api/recipes/{slug}/usage": {
      "get": {
        "description": "Used by the delete confirmation: how many meals from today onwards use this recipe.",
        "operationId": "recipe_usage",
        "parameters": [
          {
            "description": "Recipe slug, e.g. `lentil-dahl`",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Recipe slug, e.g. `lentil-dahl`",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeUsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Count upcoming planned meals",
        "tags": [
          "Recipes"
        ]
      }
    },
...
  },
...
}